// Moved rate-limit notification from backend to frontend. // Route rate limit now uses uint32, allowing max of around 42949672 flagged requests (earlier 655).
FUTURE:
// shorten log texts.
// offload logs storage to external custom servers.
// prevent mfa based activities immediately if session was revoked.

+++ Backend/processors/ratelimit/main.go
@@ -7,7 +7,7 @@ const flagKey = "RateLimitValue"
 func Set(ctx fiber.Ctx) {
 	ctx.Locals(flagKey, true)
 }
-func Get(ctx fiber.Ctx) uint16 {
+func Get(ctx fiber.Ctx) uint32 {
 	if ctx.Locals(flagKey) == true {
 		return 100
 	}
